+++ src/api/endpoints/xmltv_api.rs
@@ -5,7 +5,6 @@ use axum::response::IntoResponse;
 use chrono::{Duration, NaiveDateTime, TimeDelta};
 use flate2::write::GzEncoder;
 use flate2::Compression;
-// use actix_web::{http::header, web, HttpRequest, HttpResponse};
 use log::{error, trace};
 use quick_xml::events::{BytesStart, Event};
 use quick_xml::{Reader, Writer};