+++ src/api/model/app_state.rs
@@ -4,6 +4,7 @@ use crate::api::model::active_provider_manager::ActiveProviderManager;
 use crate::api::model::active_user_manager::ActiveUserManager;
 use crate::api::model::download::DownloadQueue;
 use crate::api::model::event_manager::EventManager;
+use crate::api::model::hls_cache::HlsCache;
 use crate::api::model::streams::shared_stream_manager::SharedStreamManager;
 use crate::model::config::{Config};
 use crate::model::hdhomerun_config::HdHomeRunDeviceConfig;
@@ -16,6 +17,7 @@ pub struct AppState {
     pub http_client: Arc<reqwest::Client>,
     pub downloads: Arc<DownloadQueue>,
     pub cache: Arc<Option<Mutex<LRUResourceCache>>>,
+    pub hls_cache: Arc<HlsCache>,
     pub shared_stream_manager: Arc<SharedStreamManager>,
     pub active_users: Arc<ActiveUserManager>,
     pub active_provider: Arc<ActiveProviderManager>,